Add(ksg) : elastic 적용중
