+++ src/main/java/com/example/meongnyangbook/config/ElasticConfig.java
@@ -1,0 +1,19 @@
+//package com.example.meongnyangbook.config;
+//
+//import co.elastic.clients.elasticsearch.ElasticsearchClient;
+//import com.example.meongnyangbook.post.repository.elastic.ElasticsearchPostRepository;
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
+//import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
+//
+//@Configuration
+//@EnableElasticsearchRepositories(basePackageClasses = {ElasticsearchPostRepository.class})
+//public class ElasticConfig {
+//
+//  @Autowired
+//  ElasticsearchOperations operations;
+//
+//  @Autowired
+//  ElasticsearchClient elasticsearchClient;
+//}
