+++ src/main/java/com/example/meongnyangbook/post/adoption/controller/AdoptionController.java
@@ -113,4 +113,11 @@ public class AdoptionController {
     AdoptionResponseDto result = adoptionService.getBestAdoptionPost();
     return ResponseEntity.status(HttpStatus.OK).body(result);
   }
+
+//  @GetMapping("/search")
+//  public ResponseEntity<List<AdoptionResponseDto>> searchAdoptionPostList(
+//      @RequestParam("keyword") String keyword) {
+//    List<AdoptionResponseDto> result = adoptionService.searchAdoptionPostList(keyword);
+//    return ResponseEntity.status(HttpStatus.OK).body(result);
+//  }
 }
