+++ src/main/java/com/example/meongnyangbook/post/adoption/repository/AdoptionSearchRepository.java
@@ -1,0 +1,5 @@
+package com.example.meongnyangbook.post.adoption.repository;
+
+public class AdoptionSearchRepository {
+
+}
