+++ src/main/java/com/example/meongnyangbook/post/entity/Post.java
@@ -4,12 +4,23 @@ import com.example.meongnyangbook.entity.TimeStamped;
 import com.example.meongnyangbook.post.attachment.entity.AttachmentUrl;
 import com.example.meongnyangbook.post.comment.entity.Comment;
 import com.example.meongnyangbook.user.entity.User;
-import jakarta.persistence.*;
-
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Inheritance;
+import jakarta.persistence.InheritanceType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.Getter;
 
+//@Document(indexName = "posts")
 @Entity
 @Getter
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS) // 변경해서 바꿀예정
