--- src/main/java/com/example/meongnyangbook/post/repository/PostRepositoryQuery.java
@@ -1,12 +1,0 @@
-package com.example.meongnyangbook.post.repository;
-
-import com.example.meongnyangbook.post.entity.Post;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface PostRepositoryQuery {
-    Post findPostById(Long postId);
-}
