+++ src/main/java/com/example/meongnyangbook/post/repository/elastic/ElasticsearchPostRepository.java
@@ -1,0 +1,6 @@
+package com.example.meongnyangbook.post.repository.elastic;
+
+//public interface ElasticsearchPostRepository extends ElasticsearchRepository<Post, Long> {
+//
+//  List<Post> findByTitleIn(Collection<String> keywords);
+//}
