+++ src/main/java/com/example/meongnyangbook/post/repository/jpa/PostRepositoryQuery.java
@@ -1,0 +1,10 @@
+package com.example.meongnyangbook.post.repository.jpa;
+
+import com.example.meongnyangbook.post.entity.Post;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepositoryQuery {
+
+  Post findPostById(Long postId);
+}
